KIWI-1568: creates /name-info route
